Refactor : SortType enums 패키지로 수정 fix : MainPageResponseDto swagger type 추가
